Move opex and capex calculation to new folder+
